Project 增加案號欄位 內帳單Excel增加日期及案號欄位 登入錯誤時顯示錯誤 web.xml移除WebSocket dependenies移除jetty7
git-svn-id: http://sanlu.googlecode.com/svn/trunk@200 bce4a620-458e-c844-f683-5f353072547c

+++ sanlu/grails-app/domain/tw/com/sanlu/Product.groovy
@@ -29,7 +29,7 @@ class Product {
 	Integer costUnit = 0
 	
 	//每次計價單位
-	Integer costRange
+	BigDecimal costRange
 	
 	//是否為場地類商品
 	boolean hasPlace
